+++ src/Simplify.Web/Bootstrapper/Setup/BaseBootstrapperPages.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using Simplify.DI;
-using Simplify.Web.Diagnostics.Measurement;
+using Simplify.Web.Diagnostics.Measurements;
 using Simplify.Web.Modules.ApplicationEnvironment;
 using Simplify.Web.Modules.Context;
 using Simplify.Web.Modules.Data;
@@ -42,13 +42,20 @@ public partial class BaseBootstrapper
 			return;
 
 		BootstrapperFactory.ContainerProvider.Register<IReadOnlyList<IPageCompositionStage>>(r =>
-			[
+		{
+			var stages = new List<IPageCompositionStage>
+			{
 				new StringTableItemsInjectionStage(r.Resolve<IStringTable>()),
 				new LanguageInjectionStage(r.Resolve<ILanguageManagerProvider>()),
 				new EnvironmentVariablesInjectionStage(r.Resolve<IDynamicEnvironment>()),
 				new ContextVariablesInjectionStage(r.Resolve<IWebContextProvider>()),
 				new SiteTitleInjectionStage(r.Resolve<IWebContextProvider>(), r.Resolve<IStringTable>()),
-				new StopwatchDataInjectionStage(r.Resolve<IStopwatchProvider>())
-			]);
+			};
+
+			if (Settings.MeasurementsEnabled)
+				stages.Add(new StopwatchDataInjectionStage(r.Resolve<IStopwatchProvider>()));
+
+			return stages;
+		});
 	}
 }