+++ src/Simplify.Web/Bootstrapper/SimplifyWebRegistrationsOverride/RegistrationsOverrideDiagnostics.cs
@@ -1,6 +1,6 @@
 using System;
 using Simplify.DI;
-using Simplify.Web.Diagnostics.Measurement;
+using Simplify.Web.Diagnostics.Measurements;
 
 namespace Simplify.Web.Bootstrapper.SimplifyWebRegistrationsOverride;